Move client creation to separate module
