+++ crates/kiwi-talk-app/src/app/mod.rs
@@ -1,10 +1,9 @@
+pub mod client;
 pub mod conn;
 pub mod constants;
 pub mod stream;
 
-use kiwi_talk_client::{
-    event::KiwiTalkClientEvent, ClientCredential, KiwiTalkClient, KiwiTalkClientEventReceiver,
-};
+use kiwi_talk_client::{event::KiwiTalkClientEvent, KiwiTalkClient, KiwiTalkClientEventReceiver};
 use serde::{Deserialize, Serialize};
 use tauri::{
     generate_handler,
@@ -16,10 +15,7 @@ use tokio::sync::{Mutex, RwLock};
 
 use crate::error::impl_tauri_error;
 
-use self::{
-    conn::{checkin, ConnError},
-    stream::{create_secure_stream, LOCO_CLIENT_SECURE_SESSION},
-};
+use self::client::{create_client, CreateClientError};
 
 pub fn init_plugin<R: Runtime>(name: &'static str) -> TauriPlugin<R> {
     Builder::new(name)
@@ -32,6 +28,14 @@ pub fn init_plugin<R: Runtime>(name: &'static str) -> TauriPlugin<R> {
         .build()
 }
 
+#[derive(Default)]
+struct KiwiTalkApp {
+    pub credential: RwLock<Option<AppCredential>>,
+
+    pub client: RwLock<Option<KiwiTalkClient>>,
+    pub client_events: Mutex<Option<KiwiTalkClientEventReceiver>>,
+}
+
 fn setup_plugin<R: Runtime>(handle: &AppHandle<R>) -> tauri::plugin::Result<()> {
     handle.manage(KiwiTalkApp::default());
 
@@ -39,18 +43,10 @@ fn setup_plugin<R: Runtime>(handle: &AppHandle<R>) -> tauri::plugin::Result<()> 
 }
 
 #[derive(Serialize, Deserialize)]
-struct AppCredential {
+pub struct AppCredential {
     pub access_token: String,
     pub refresh_token: String,
     pub user_id: Option<i64>,
-}
-
-#[derive(Default)]
-struct KiwiTalkApp {
-    pub credential: RwLock<Option<AppCredential>>,
-
-    pub client: RwLock<Option<KiwiTalkClient>>,
-    pub client_events: Mutex<Option<KiwiTalkClientEventReceiver>>,
 }
 
 #[tauri::command(async)]
@@ -69,14 +65,8 @@ pub enum ClientInitializeError {
     #[error("Credential is not set")]
     CredentialNotSet,
 
-    #[error("Checkin failed")]
-    Checkin,
-
-    #[error("Loco stream handshake failed")]
-    LocoHandshake,
-
-    #[error("Error while initializing client")]
-    Client,
+    #[error("Client creation failed")]
+    Client(#[from] CreateClientError),
 }
 
 impl_tauri_error!(ClientInitializeError);
@@ -88,31 +78,7 @@ async fn initialize_client(app: State<'_, KiwiTalkApp>) -> Result<(), ClientInit
             let mut client_slot = app.client.write().await;
             let mut client_events_slot = app.client_events.lock().await;
 
-            let checkin_res = checkin(credential.user_id.unwrap_or(1))
-                .await
-                .map_err(|_| ClientInitializeError::Checkin)?;
-            if checkin_res.status != 0 || checkin_res.data.is_none() {
-                return Err(ClientInitializeError::Checkin);
-            }
-            let checkin_res = checkin_res.data.unwrap();
-
-            let loco_session = create_secure_stream(
-                &LOCO_CLIENT_SECURE_SESSION,
-                (checkin_res.host.as_str(), checkin_res.port as u16),
-            )
-            .await
-            .map_err(|_| ClientInitializeError::LocoHandshake)?;
-
-            let (client, client_events) = KiwiTalkClient::login(
-                loco_session,
-                ClientCredential {
-                    access_token: &credential.access_token,
-                    device_uuid: crate::auth::constants::DEVICE_UUID, // TODO:: REPLACE
-                    user_id: credential.user_id,
-                },
-            )
-            .await
-            .map_err(|_| ClientInitializeError::Client)?;
+            let (client, client_events) = create_client(credential).await?;
 
             *client_slot = Some(client);
             *client_events_slot = Some(client_events);
